Fix dynamic zone modification, add service
If you modify a dynamic zone and have a higher serial, bind will refuse
to load the zone. We fix this by freezing the zone before the
modification and reloading it after.

Warning: this erases all 'dynamic' modifications

To not lose all modifications, you have to set the serial to the latest
one that chef used. Be sure to upgrade it after by any nslookup addition
so the slaves can import your latest modification.

Set the service and enable/start it. Currently, no reload on
configuration changes.