Fix how to guess zones, use chef.kitchen for tests
Change the way zones are determined from domains. It was from the last
element of the domain but it could not work apart for strange zones
like 'kitchen'. Now, it takes the whole tail so 'a.b.c' gives 'b.c'
(instead of only 'c' with previous method).

Move test zone from 'kitchen' to 'chef.kitchen' to be able to test this
new behaviour.