Fix resolve_iface when get ip during current run
Add a ruby_block which set the ip address of an interface during
convergence if it could not get it during compilation, probably because
it was not set at that moment.

Reimplement resolve_iface in native ruby with socket libraby:
- Previous implementation of resolve_iface was relying on Ohai. So it
was not updated if a network modification occured during the chef run.
This caused problems when it was used with dhcp where ip address was
fetched during the same run.
- This new implementation uses socket library to get the current network
state, not the one from the beginning of the chef run.