Eval input in resolve_iface to allow attributes
Move evaluation of record data from nsupdate to default library in
resolve_iface to be more generic. For instance, this allows the
'a' field of a zone to be configured with
node['network']['default_interface'].

Warning: this evaluation is unsafe (although being able to configure
chef attributes should be considered as equivalent as root access),
it should probably be replaced by a safer implementation.